This one is gonna be big.
Ive converted my entire backend into something a lot more flexible imo. Instead of using an eternal Postgres DB, DrizzleORM, and astro actions, I have opted for PGLite, TypeORM, and TRPC. This should allow me to ensure a seamless backend that is capable of notifying the frontend of changes in data.
additionally im trying out snyk for security scanning so there are a couple of files for that.