using rails in a way that works for me
as much as i currently love rails for its design and ease of use, it
doesnt exactly handle templating in a way that makes a ton of sense to
my barin. as such i pulled in phlex to allow for easy components inside
of rails classes. unfortunately this means i need to rebuild some of the
basic structure, but this should allow me to work faster overall.